requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
astunparse=1.6.3=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.3.2=pypi_0
cloudpickle=3.0.0=pypi_0
decorator=5.1.1=pypi_0
dm-tree=0.1.8=pypi_0
filelock=3.15.4=pypi_0
flatbuffers=24.3.25=pypi_0
fsspec=2024.9.0=pypi_0
gast=0.6.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.66.1=pypi_0
h5py=3.11.0=pypi_0
huggingface-hub=0.24.6=pypi_0
idna=3.8=pypi_0
keras=3.5.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=18.1.1=pypi_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.4.0=pypi_0
mpmath=1.3.0=pypi_0
namex=0.0.8=pypi_0
ncurses=6.4=h6a678d5_0
nose=1.3.7=pypi_0
numpy=1.26.4=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.15=h5eee18b_0
opt-einsum=3.3.0=pypi_0
optree=0.12.1=pypi_0
packaging=24.1=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py310h06a4308_0
protobuf=4.25.4=pypi_0
ptflops=0.7.3=pypi_0
pygments=2.18.0=pypi_0
python=3.10.14=h955ad1f_1
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
rich=13.8.0=pypi_0
robust-loss-pytorch=0.0.2=dev_0
safetensors=0.4.5=pypi_0
scipy=1.14.1=pypi_0
setuptools=72.1.0=py310h06a4308_0
six=1.16.0=pypi_0
sqlite=3.45.3=h5eee18b_0
tensorboard=2.17.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.17.0=pypi_0
tensorflow-io-gcs-filesystem=0.37.1=pypi_0
tensorflow-probability=0.24.0=pypi_0
termcolor=2.4.0=pypi_0
timm=1.0.9=pypi_0
tk=8.6.14=h39e8969_0
torch=1.13.0+cu116=pypi_0
torch-dct=0.1.6=pypi_0
torchaudio=0.13.0+cu116=pypi_0
torchvision=0.14.0+cu116=pypi_0
tqdm=4.66.5=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=pypi_0
werkzeug=3.0.4=pypi_0
wheel=0.43.0=py310h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
zlib=1.2.13=h5eee18b_1
